освещение и не только...

+++ core/src/com/gungame/world/objects/weapon/BulletData.java
@@ -1,5 +1,5 @@
 package com.gungame.world.objects.weapon;
 
-public record BulletData(int speed, float deviation, int damage, short shotID) {
+public record BulletData(float speed, float deviation, int damage, short shotID) {
 
 }
